# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bsllmner-viewer API",
    "description": "Queries, aggregations, and exports over ontology-mapped BioSample annotations. The conventions follow the DDBJ Search API.\n\nStart with [/llms.txt](/llms.txt): condition examples and recipes for common tasks. [/llms-full.txt](/llms-full.txt) has the rules that span operations: the condition language, the descendants of terms, counting units and populations, self-exclusion, default elements, expected counts, export formats, errors, limits, and caching with `ETag` and `If-None-Match`. A description in this document names the heading of the rule that it relies on.",
    "contact": {
      "name": "BioData Science Initiative",
      "url": "https://github.com/dbcls/bsllmner-viewer"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/dataset": {
      "get": {
        "tags": [
          "Dataset"
        ],
        "summary": "Get the dataset version, the fields, and the population totals",
        "description": "Returns what a client needs before it writes conditions: the version of the dataset, the target assays, the fields of the condition language with their kinds, the status groups, and the BioSample counts of the whole population per assay, organism, and field. Counts are in the BioSample unit, except `totals`, which has all three units.",
        "operationId": "getDataset",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/distribution": {
      "get": {
        "tags": [
          "Aggregations"
        ],
        "summary": "Count the population per element of one dimension",
        "description": "Counts the population per element of the dimension `field`, in `unit`. The population is `q`, or `q` without the conjuncts on `field` with `facetSelfExclude`. `total` is the count of the population. An element counts the units that have its value. For a term dimension, the units have the term or one of its descendants. The Distribution view of the UI calls this operation once per field. See \"Aggregations\" and \"Default elements\" in /llms-full.txt for the elements and their order.",
        "operationId": "getDistribution",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "description": "Dimension of the counts. An annotation term field, `library_strategy`, `organism_id`, or `date_published`. `dslFields` of `GET /api/dataset` lists the fields with their kinds `term`, `assay`, `organism`, and `date`",
              "title": "Field"
            },
            "description": "Dimension of the counts. An annotation term field, `library_strategy`, `organism_id`, or `date_published`. `dslFields` of `GET /api/dataset` lists the fields with their kinds `term`, `assay`, `organism`, and `date`"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          },
          {
            "name": "unit",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Unit",
              "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt",
              "default": "biosample"
            },
            "description": "Counting unit. `biosample` counts distinct BioSamples, `sra-experiment` distinct SRA Experiments, and `bioproject` distinct BioProjects. See \"Counting\" in /llms-full.txt"
          },
          {
            "name": "facetSelfExclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt",
              "default": false,
              "title": "Facetselfexclude"
            },
            "description": "If true, the operation removes from `q` each top-level conjunct whose clauses are all on the dimensions of the operation, as the UI does. The top-level conjuncts are the operands of the `AND` at the root of `q`, or `q` itself if the root is not `AND`. A conjunct that has a keyword stays. The description of the operation names its dimensions. If false, the operation uses `q` as it is. See \"Self-exclusion\" in /llms-full.txt"
          },
          {
            "name": "elements",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated elements, at most 100. An element is a term ID for an annotation term field, a target assay for `library_strategy`, an NCBI Taxonomy ID for `organism_id`, and a year for `date_published`. See \"Default elements\" in /llms-full.txt",
              "title": "Elements"
            },
            "description": "Comma-separated elements, at most 100. An element is a term ID for an annotation term field, a target assay for `library_strategy`, an NCBI Taxonomy ID for `organism_id`, and a year for `date_published`. See \"Default elements\" in /llms-full.txt"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Number of default elements, from 1 to 200, when `elements` is omitted. The elements that `q` names come in addition. It does not apply to `date_published`, which has every year",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of default elements, from 1 to 200, when `elements` is omitted. The elements that `q` names come in addition. It does not apply to `date_published`, which has every year"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.\n- `invalid-dimension`: A field exists but cannot be the dimension of the operation, or the two dimensions of a cross-tabulation are the same field. Choose from the fields that `detail` lists.\n- `invalid-element`: A named element does not fit its dimension, such as an organism that is not an NCBI Taxonomy ID or a year that is not a number. Name elements of the form that `detail` gives, or omit them to get the default elements.\n- `too-many-elements`: A request names more elements of one dimension than the limit of 100. Name at most 100 elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetAssay": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Target assay, as a `library_strategy` value"
          },
          "biosampleCount": {
            "type": "integer",
            "title": "Biosamplecount",
            "description": "BioSamples of the whole population with an SRA Experiment of the assay"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "biosampleCount"
        ],
        "title": "DatasetAssay"
      },
...
      "DatasetResponse": {
        "properties": {
          "datasetVersion": {
            "$ref": "#/components/schemas/DatasetVersionRef"
          },
          "version": {
            "additionalProperties": true,
            "type": "object",
            "title": "Version",
            "description": "Full dataset version information"
          },
          "targetAssays": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Targetassays",
            "description": "Target assays of the dataset. They are the values that `library_strategy` accepts in a condition"
          },
          "assays": {
            "items": {
              "$ref": "#/components/schemas/DatasetAssay"
            },
            "type": "array",
            "title": "Assays",
            "description": "Target assays in descending order of their BioSamples"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldDescription"
            },
            "type": "array",
            "title": "Fields",
            "description": "Annotation fields, in the order of the select configuration"
          },
          "dslFields": {
            "items": {
              "$ref": "#/components/schemas/DslFieldDescription"
            },
            "type": "array",
            "title": "Dslfields",
            "description": "Every field that a condition can name"
          },
          "statuses": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Status"
              },
              "type": "array"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/GroupName"
            },
            "type": "object",
            "title": "Statuses",
            "description": "Status groups and the statuses under them. A condition names a group. An entry and an export report the status of each annotation"
          },
          "totals": {
            "$ref": "#/components/schemas/Totals",
            "description": "Counts of the whole population in each counting unit"
          },
          "organisms": {
            "items": {
              "$ref": "#/components/schemas/DatasetOrganism"
            },
            "type": "array",
            "title": "Organisms",
            "description": "Organisms of the population in descending order of their BioSamples"
          },
          "ontologies": {
            "items": {
              "$ref": "#/components/schemas/DatasetOntology"
            },
            "type": "array",
            "title": "Ontologies",
            "description": "Names of the prefixes of the terms of the dataset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "datasetVersion",
          "version",
          "targetAssays",
          "assays",
          "fields",
          "dslFields",
          "statuses",
          "totals",
          "organisms",
          "ontologies"
        ],
        "title": "DatasetResponse"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "`about:blank` for an error that the HTTP status describes. Otherwise a URI that ends with the slug of the rule that the request broke. The description of the 400 response of each operation lists its slugs."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The HTTP status phrase"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "The path of the request"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 time in UTC"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid",
            "description": "The value of the `X-Request-ID` header of the response"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "timestamp",
          "requestId"
        ],
        "title": "ProblemDetails",
        "description": "RFC 7807 problem details of an error response."
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "mapped_exact",
          "mapped_selected",
          "unmapped_no_candidate",
          "unmapped_rejected",
          "not_stated",
          "extraction_failed"
        ]
      },
...
      "Unit": {
        "type": "string",
        "enum": [
          "biosample",
          "sra-experiment",
          "bioproject"
        ]
      }
    }
  },
...
}